Break up Canvas class
