+++ spec/suitec/canvas_assignment_sync_spec.rb
@@ -12,7 +12,7 @@ describe 'Canvas assignment sync', order: :defined do
     @course = Course.new({})
     @course.site_id = course_id
     @driver = Utils.launch_browser
-    @canvas = Page::CanvasPage.new @driver
+    @canvas = Page::CanvasActivitiesPage.new @driver
     @cal_net = Page::CalNetPage.new @driver
     @asset_library = Page::SuiteCPages::AssetLibraryPage.new @driver
     @engagement_index = Page::SuiteCPages::EngagementIndexPage.new @driver
