+++ util/spec_helper.rb
@@ -22,7 +22,8 @@ require_relative '../logging'
 require_relative 'utils'
 require_relative '../pages/page'
 require_relative '../pages/cal_net_page'
-require_relative '../pages/canvas_page'
+require_relative '../pages/canvas/canvas_page'
+require_relative '../pages/canvas/canvas_activities_page'
 require_relative '../pages/junction/junction_pages'
 require_relative '../pages/junction/api_academics_course_provision_page'
 require_relative '../pages/junction/api_academics_roster_page'
